Move ensureHttps from R4ImportBundleProducer to Uris Utility Class
