+++ cqf-fhir-cr/src/main/java/org/opencds/cqf/fhir/cr/ecr/r4/R4ImportBundleProducer.java
@@ -4,9 +4,7 @@ import ca.uhn.fhir.context.FhirVersionEnum;
 import ca.uhn.fhir.repository.IRepository;
 import ca.uhn.fhir.rest.server.exceptions.UnprocessableEntityException;
 import java.net.MalformedURLException;
-import java.net.URI;
 import java.net.URISyntaxException;
-import java.net.URL;
 import java.util.*;
 import java.util.function.Function;
 import java.util.function.Predicate;
@@ -15,6 +13,7 @@ import org.hl7.fhir.instance.model.api.IBaseResource;
 import org.hl7.fhir.r4.model.*;
 import org.opencds.cqf.fhir.cr.crmi.TransformProperties;
 import org.opencds.cqf.fhir.cr.ecr.FhirResourceExistsException;
+import org.opencds.cqf.fhir.utility.Uris;
 import org.opencds.cqf.fhir.utility.adapter.IAdapterFactory;
 import org.opencds.cqf.fhir.utility.adapter.IKnowledgeArtifactAdapter;
 import org.opencds.cqf.fhir.utility.search.Searches;
@@ -92,27 +91,6 @@ public class R4ImportBundleProducer {
             ext.setValue(new UriType(url));
             vs.getExtension().add(ext);
         }
-    }
-
-    public static String ensureHttps(String urlString) throws MalformedURLException, URISyntaxException {
-        URL url = URI.create(urlString).toURL();
-
-        // Check if the protocol is already HTTPS
-        if ("https".equalsIgnoreCase(url.getProtocol())) {
-            return urlString;
-        }
-
-        // Construct a new URL with the HTTPS protocol
-        URI httpsUrl = new URI(
-                "https", // scheme
-                null, // userinfo
-                url.getHost(), // host
-                url.getPort(), // port
-                url.getFile(), // path
-                null, // query
-                null // fragment
-                );
-        return httpsUrl.toString();
     }
 
     public static List<Bundle.BundleEntryComponent> transformImportBundle(
@@ -227,7 +205,7 @@ public class R4ImportBundleProducer {
         String valueSetAuthoritativeSourceUrl = valueSet.getUrl();
 
         try {
-            valueSetAuthoritativeSourceUrl = ensureHttps(valueSetAuthoritativeSourceUrl);
+            valueSetAuthoritativeSourceUrl = Uris.ensureHttps(valueSetAuthoritativeSourceUrl);
         } catch (URISyntaxException | MalformedURLException e) {
             // Do nothing here and let the malformed URL flow through.
         }
