+++ cqf-fhir-cr/src/main/java/org/opencds/cqf/fhir/cr/visitor/dstu3/ReleaseVisitor.java
@@ -30,12 +30,12 @@ import org.hl7.fhir.instance.model.api.IBaseParameters;
 import org.hl7.fhir.instance.model.api.IIdType;
 import org.hl7.fhir.instance.model.api.IPrimitiveType;
 import org.opencds.cqf.fhir.cr.crmi.TransformProperties;
-import org.opencds.cqf.fhir.cr.ecr.r4.R4ImportBundleProducer;
 import org.opencds.cqf.fhir.cr.visitor.dstu3.CRMIReleaseExperimentalBehavior.CRMIReleaseExperimentalBehaviorCodes;
 import org.opencds.cqf.fhir.cr.visitor.dstu3.CRMIReleaseVersionBehavior.CRMIReleaseVersionBehaviorCodes;
 import org.opencds.cqf.fhir.utility.Constants;
 import org.opencds.cqf.fhir.utility.PackageHelper;
 import org.opencds.cqf.fhir.utility.SearchHelper;
+import org.opencds.cqf.fhir.utility.Uris;
 import org.opencds.cqf.fhir.utility.adapter.IKnowledgeArtifactAdapter;
 import org.opencds.cqf.fhir.utility.dstu3.ArtifactAssessment;
 import org.slf4j.Logger;
@@ -187,7 +187,7 @@ public class ReleaseVisitor {
 
     public static void addAuthoritativeSourceExtension(ValueSet valueSet, String url) {
         try {
-            url = R4ImportBundleProducer.ensureHttps(url);
+            url = Uris.ensureHttps(url);
         } catch (URISyntaxException | MalformedURLException e) {
             // Do nothing here and let the malformed URL flow through.
         }
