+++ cqf-fhir-utility/src/main/java/org/opencds/cqf/fhir/utility/Uris.java
@@ -1,5 +1,10 @@
 package org.opencds.cqf.fhir.utility;
 
+import java.net.MalformedURLException;
+import java.net.URI;
+import java.net.URISyntaxException;
+import java.net.URL;
+
 public class Uris {
     private Uris() {}
 
@@ -17,5 +22,26 @@ public class Uris {
         }
 
         return uri.startsWith("file") || !uri.matches("\\w+?://.*");
+    }
+
+    public static String ensureHttps(String urlString) throws MalformedURLException, URISyntaxException {
+        URL url = URI.create(urlString).toURL();
+
+        // Check if the protocol is already HTTPS
+        if ("https".equalsIgnoreCase(url.getProtocol())) {
+            return urlString;
+        }
+
+        // Construct a new URL with the HTTPS protocol
+        URI httpsUrl = new URI(
+                "https", // scheme
+                null, // userinfo
+                url.getHost(), // host
+                url.getPort(), // port
+                url.getFile(), // path
+                null, // query
+                null // fragment
+                );
+        return httpsUrl.toString();
     }
 }
